scripts/twitter-scraper/requirements.txt
requests
beautifulsoup4
python-dotenv
openai
kor
langchain
langchain-community
langchain-openai
